# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:52075","description":"Generated server url"}],"paths":{"/update-car/{id}":{"put":{"tags":["car-controller"],"operationId":"updateCar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CarResponseDto"}}}}}}},"/rents/update-rent/{id}":{"put":{"tags":["rent-controller"],"operationId":"updateRent","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RentRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RentResponseDto"}}}}}}},"/car-type/update/{id}":{"put":{"tags":["car-type-controller"],"operationId":"updateCar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarTypeRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CarTypeResponseDto"}}}}}}},"/rents/create-rent":{"post":{"tags":["rent-controller"],"operationId":"createRent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RentRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RentResponseDto"}}}}}}},"/create-car":{"post":{"tags":["car-controller"],"operationId":"createCar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CarResponseDto"}}}}}}},"/car-type":{"post":{"tags":["car-type-controller"],"operationId":"createCarType","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarTypeRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CarTypeResponseDto"}}}}}}},"/rents/get-rent-by-id/{id}":{"get":{"tags":["rent-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RentResponseDto"}}}}}}},"/rents/get-all":{"get":{"tags":["rent-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RentResponseDto"}}}}}}}},"/get-car/{id}":{"get":{"tags":["car-controller"],"operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CarResponseDto"}}}}}}},"/get-all":{"get":{"tags":["car-controller"],"operationId":"getAllCars","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CarResponseDto"}}}}}}}},"/car-type/{id}":{"get":{"tags":["car-type-controller"],"operationId":"getById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CarTypeResponseDto"}}}}}}},"/car-type/get-all":{"get":{"tags":["car-type-controller"],"operationId":"getAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CarTypeResponseDto"}}}}}}}},"/Pong":{"get":{"tags":["ping-controller"],"operationId":"Ping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/rents/delete/{id}":{"delete":{"tags":["rent-controller"],"operationId":"deleteRent","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/delete-car/{id}":{"delete":{"tags":["car-controller"],"operationId":"deleteCar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/car-type/delete/{id}":{"delete":{"tags":["car-type-controller"],"operationId":"deleteCarType","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"CarRequestDto":{"type":"object","properties":{"brand":{"type":"string"},"model":{"type":"string"},"idcarType":{"type":"integer","format":"int64"}}},"CarResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"carType":{"$ref":"#/components/schemas/CarType"},"brand":{"type":"string"},"model":{"type":"string"}}},"CarType":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string"},"price":{"type":"number"}}},"RentRequestDto":{"type":"object","properties":{"customer":{"type":"string"},"idCar":{"type":"integer","format":"int64"},"rentedDays":{"type":"integer","format":"int32"},"startRent":{"type":"string","format":"date-time"},"endRent":{"type":"string","format":"date-time"},"totalPrice":{"type":"number"}}},"Car":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"carType":{"$ref":"#/components/schemas/CarType"},"brand":{"type":"string"},"model":{"type":"string"}}},"RentResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"customer":{"type":"string"},"car":{"$ref":"#/components/schemas/Car"},"rentedDays":{"type":"integer","format":"int32"},"startRent":{"type":"string","format":"date-time"},"endRent":{"type":"string","format":"date-time"},"totalPrice":{"type":"number"}}},"CarTypeRequestDto":{"type":"object","properties":{"type":{"type":"string"},"price":{"type":"number"}}},"CarTypeResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string"},"price":{"type":"number"}}}}}}